R moved examples out of src directory
