+++ .eslintrc
@@ -4,7 +4,7 @@
   ],
   "overrides": [
     {
-      "files": ["**/*.test.ts"],
+      "files": ["**/*.test.ts", "examples/*.ts"],
       "parserOptions": {
         "project": "./tsconfig.eslint.json"
       },
