+++ examples/exec.ts
@@ -1,8 +1,8 @@
 import { execSync } from 'child_process';
-import { CliStringInput } from '../cli-string-input';
-import { CliLeaf } from '../cli-leaf';
-import { CliStringArrayInput } from '../cli-string-array-input';
-import { runCliAndExit } from '../run-cli-and-exit';
+import { CliStringInput } from '../src/cli-string-input';
+import { CliLeaf } from '../src/cli-leaf';
+import { CliStringArrayInput } from '../src/cli-string-array-input';
+import { runCliAndExit } from '../src/run-cli-and-exit';
 
 export const execCliLeaf = CliLeaf({
   name: 'exec',
