+++ tsconfig.eslint.json
@@ -1,5 +1,5 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./src", "./tests"],
+  "include": ["./src", "./tests", "./examples"],
   "compilerOptions": {}
 }
